Split exchange buffer files into two, fixed text template formatting and fixed some incorrect comments
