+++ Src/ILGPU/ILGPU.csproj
@@ -80,10 +80,10 @@
           <AutoGen>True</AutoGen>
           <DependentUpon>IntrinsicMatchers.tt</DependentUpon>
         </None>
-        <None Include="Runtime\ExchangeBuffer.cs">
+        <None Include="Runtime\ExchangeBufferExtensions.cs">
           <DesignTime>True</DesignTime>
           <AutoGen>True</AutoGen>
-          <DependentUpon>ExchangeBuffer.tt</DependentUpon>
+          <DependentUpon>ExchangeBufferExtensions.tt</DependentUpon>
         </None>
         <None Include="Runtime\KernelLoaders.cs">
           <DesignTime>True</DesignTime>
@@ -163,6 +163,11 @@
           <AutoGen>True</AutoGen>
           <DependentUpon>ExchangeBuffer.tt</DependentUpon>
         </Compile>
+        <Compile Update="Runtime\ExchangeBufferExtensions.cs">
+          <DesignTime>True</DesignTime>
+          <AutoGen>True</AutoGen>
+          <DependentUpon>ExchangeBufferExtensions.tt</DependentUpon>
+        </Compile>
         <Compile Update="Runtime\KernelLoaders.cs">
           <DesignTime>True</DesignTime>
           <AutoGen>True</AutoGen>
@@ -238,9 +243,9 @@
           <Generator>TextTemplatingFileGenerator</Generator>
           <LastGenOutput>IntrinsicMatchers.cs</LastGenOutput>
         </None>
-        <None Update="Runtime\ExchangeBuffer.tt">
+        <None Update="Runtime\ExchangeBufferExtensions.tt">
           <Generator>TextTemplatingFileGenerator</Generator>
-          <LastGenOutput>ExchangeBuffer.cs</LastGenOutput>
+          <LastGenOutput>ExchangeBufferExtensions.cs</LastGenOutput>
         </None>
         <None Update="Runtime\KernelLoaders.tt">
           <Generator>TextTemplatingFileGenerator</Generator>
